refactor(deployments): split payload builders
